refactor entire horizontal list component to create cleaner file structure with greater ease of reuse
