+++ src/main.tsx
@@ -3,7 +3,11 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import "./global/default.css";
-import App from "../src/components/App/App";
+// import App from "../src/components/App/App";
+// import { HorizontalListImage } from "./components/HorizontalList/HorizontalListImage/HorizontalListImage";
+// import { HorizontalListHome } from "./components/HorizontalList/HorizontalListHome/HorizontalListHome";
+// import { HorizontalListStaffing } from "./components/HorizontalList/HorizontalListStaffing/HorizontalListStaffing";
+
 const client = new ApolloClient({
   uri: "http://localhost:8080/api/graphql",
   cache: new InMemoryCache(),
@@ -13,7 +17,10 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <ApolloProvider client={client}>
       <BrowserRouter>
-        <App />
+        {/* <HorizontalListImage /> */}
+        {/* <HorizontalListHome /> */}
+        {/* <HorizontalListStaffing /> */}
+        {/* <App /> */}
       </BrowserRouter>
     </ApolloProvider>
   </React.StrictMode>,
